Uploaded files for Project
There is a JSON file that basically contains the world map.
The VOlcanoes_USA contains the coordinates of some volcanoes in the USA.
The map1.py reads the text file and puts markers on where the volcanos are. The colours depend on the elevation of the volcano. Also if a country is coloured green, it has less than 10 million people, yellow is above 10 million but less than 20 million, and red is above 20 million